backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
reportlab==4.0.7
openpyxl==3.1.2
geopy==2.4.1
gunicorn==21.2.0
flask-mongoengine==1.0.0

